+++ .storybook/config.js
@@ -2,7 +2,6 @@ import { configure } from '@kadira/storybook';
 import '../src/index.css';
 
 function loadStories() {
-  require('../src/containers/stories/');
   require('../src/components/stories/');
 }
