+++ src/components/stories/index.js
@@ -1,3 +1,4 @@
 import './Card';
 import './Icon';
 import './Toast';
+import './FindBar';
