--- src/containers/stories/index.js
@@ -1,1 +1,0 @@
-import './FindBar';
